Miscellaneous code cleanups
- Drop STANDALONE_SERVER option for now, as it is not functional
- Remove unnecessary header includes
- Drop old duplicate DEBUGLOG macro and its backend function
- Other miscellaneous fixed and style changes

+++ include/resources/af.h
@@ -5,7 +5,7 @@
 
 typedef struct af_t {
     unsigned int id;
-    unsigned int endurance;
+    float endurance;
     unsigned int health;
     float forward_speed;
     float reverse_speed;
